==> 2_distribution_impacts.R <==
##R code for the ecological analysis of our review
##Author: Alba Aguion, Iratxe Rubio
##Date: August 20th 2017


#libraries
library(rockchalk)
library(tidyverse)
library(reshape2) #melt and dcast function
library(cowplot)


#open dataset
data <- read.table("exploited_ccshifts_database.csv", header= T, sep=",")

#Factors
data$id_study <- as.factor(data$id_study)
data$id_obs <- as.factor(data$id_obs)
data$year_publication <- as.factor(data$year_publication)


#Numeric
data$years_data <- as.numeric(as.character(data$years_data), na.omit = T)


##############################################
# Figure CLIMATE CHANGE VARIABLES PER IMPACT #
##############################################

counts <- dcast(data, response ~ cc_driver_detail, fun.aggregate = length)

cc_driver.counts <- gather(counts,"cc_driver", "counts", 2:8)

ggplot(cc_driver.counts, aes(x = response, y= counts , fill = cc_driver)) + 
  geom_bar(stat="identity", alpha=0.8) + 
  scale_fill_manual(values = c("darkgreen", "grey", "lightblue", 
                               "yellow", "deeppink4", "black", "red")) +
  labs(y = "Number of Observations", 
       x = "", 
       fill = "Climate Change variables") + 
  theme(panel.background = element_rect(fill = 'whitesmoke', 
                                        colour = 'black'), 
        legend.position= "left") 


############################
# Figure IMPACTS MAGNITUDE #
############################

############
# LATITUDE #
############

#histogram
lat <- subset (data, response == "latitude")
lat$sign <- ifelse(lat$decadal_change > 0,"Polewards (+)", "Towards equator (-)") #there are no zeros sum(lat$decadal_change == 0)

Fig2.lat <- lat %>% 
  ggplot() +
  geom_histogram(aes(decadal_change, fill = sign, colour = sign),
                 alpha = 0.7, binwidth = 9) +
  scale_x_continuous(name = "Mean Latitude Shift Rate (km/decade)") + 
  scale_y_continuous(name = "Number of Observations") +
  scale_colour_manual(name = "Shift direction", 
                      values = c("Polewards (+)" = "skyblue", 
                                 "Towards equator (-)" = "royalblue4"), 
                      labels = c("Polewards (+)" = "Polewards", 
                                 "Towards equator (-)" = "Towards equator")) +
  scale_fill_manual(name = "Shift direction", 
                    values = c("Polewards (+)" = "skyblue", 
                               "Towards equator (-)"="royalblue4"), 
                    labels = c("Polewards (+)" = "Polewards", 
                               "Towards equator (-)" = "Towards equator")) +
  theme(panel.background = element_rect(fill = 'whitesmoke',
                                        colour = 'black'),
        legend.position = c(.85, .75)) +
  coord_flip() +
  theme_bw() 

#taxa barplot
my.labels2 <- c("Benthic \n crustacea",
                "Benthic \n mollusca",
                "Bony-fish", 
                "Cephalopode",
                "Non-bony \n fish")

lat.barplot <- ggplot(lat, aes(taxa, decadal_change, fill = taxa)) + 
  geom_boxplot(data = subset(lat, decadal_change >= 0), 
               aes(taxa, decadal_change),  
               na.rm = T, outlier.shape = 1, outlier.size = 0.1) +
  geom_boxplot(data = subset(lat, decadal_change < 0), 
               aes(taxa, decadal_change),  
               na.rm = T, outlier.shape = 1, outlier.size = 0.1) +
  geom_hline(yintercept = c(0), linetype = "dotted") +
  scale_y_continuous(name = "km/decade", breaks = seq(-100, 200, by = 50)) +
  xlab(NULL)+
  theme(axis.text.x =element_blank()) +
  scale_fill_brewer(palette = "Blues") +
  theme_bw() +
  scale_x_discrete(labels = my.labels2) +
  theme(legend.position="none")

#########
# DEPTH #
#########
depth <- subset (data, response == "depth")
depth$sign <- ifelse(depth$decadal_change < 0, "Shallower (-)", "Deeper (+)")# there are no zeros sum(depth$decadal_change == 0)

depth$decadal_change_reversed <- (-1)*depth$decadal_change

Fig2.depth <- depth %>%
  ggplot()+
  geom_histogram(aes(decadal_change_reversed, fill = sign, colour = sign), 
                 alpha = 0.7, binwidth = 4) +
  scale_x_continuous(name = "Depth Shift Rate (m/decade)") + 
  scale_y_continuous(name = "Number of Observations")+
  scale_colour_manual(name = "Shift direction", 
                      values = c("Deeper (+)" = "royalblue4", 
                                 "Shallower (-)"="skyblue"), 
                      labels = c("Deeper (+)" = "Deeper", 
                                 "Shallower (-)" = "Shallower"),
                      guide = guide_legend(reverse = T)) +
  scale_fill_manual(name = "Shift direction", 
                    values = c("Deeper (+)" = "royalblue4", 
                               "Shallower (-)" = "skyblue"), 
                    labels = c("Deeper (+)" = "Deeper", 
                               "Shallower (-)" = "Shallower"),
                    guide = guide_legend(reverse = T)) +
  theme(panel.background = element_rect(fill = 'whitesmoke', 
                                        colour = 'black'),
        legend.position = c(.85, .75)) +
  coord_flip() +
  theme_bw()

#taxa barplot depth
depth.barplot <- ggplot(depth, aes(taxa, decadal_change, fill = taxa)) +
  geom_boxplot(data = subset(depth, decadal_change >= 0), 
               aes(taxa, decadal_change),  
               na.rm = T, outlier.shape = 1, outlier.size = 0.1) +
  geom_boxplot(data = subset(depth, decadal_change < 0), 
               aes(taxa, decadal_change),  
               na.rm = T, outlier.shape = 1, outlier.size = 0.1) +
  geom_hline(yintercept = c(0), linetype = "dotted")+
  scale_y_continuous(name ="m/decade", breaks = seq(-80, 60, by = 20)) +
  xlab(NULL) +
  theme(axis.text.x = element_blank()) +
  scale_fill_brewer(palette = "Blues") +
  theme_bw() +
  scale_x_discrete(labels = my.labels2) +
  theme(legend.position = "none")

join.lat.depth <- plot_grid(lat.barplot, Fig2.lat, depth.barplot, Fig2.depth, labels = c("A", "B", "C", "D"), align="hv")
plot(join.lat.depth)
#ggsave("join_lat_depth.jpeg")
